--- a/core/java/android/app/AppOpsManager.java
+++ b/core/java/android/app/AppOpsManager.java
@@ -28,6 +28,7 @@ import android.annotation.RequiresPermission;
 import android.annotation.SystemApi;
 import android.annotation.SystemService;
 import android.annotation.TestApi;
+import android.app.compat.gms.GmsCompat;
 import android.app.usage.UsageStatsManager;
 import android.compat.Compatibility;
 import android.compat.annotation.ChangeId;
@@ -8263,6 +8264,10 @@ public class AppOpsManager {
      */
     public int noteOpNoThrow(int op, int uid, @Nullable String packageName,
             @Nullable String attributionTag, @Nullable String message) {
+        if (GmsCompat.isEnabled() && uid != Process.myUid()) {
+            return noteProxyOpNoThrow(op, packageName, uid, attributionTag, message);
+        }
+
         try {
             collectNoteOpCallsForValidation(op);
             int collectionMode = getNotedOpCollectionMode(uid, packageName, op);
@@ -8410,6 +8415,14 @@ public class AppOpsManager {
                 mContext.getAttributionSource(), new AttributionSource(proxiedUid,
                         proxiedPackageName, proxiedAttributionTag, mContext.getAttributionSource()
                         .getToken())), message,/*skipProxyOperation*/ false);
+    }
+
+    private int noteProxyOpNoThrow(int op, @Nullable String proxiedPackageName,
+            int proxiedUid, @Nullable String proxiedAttributionTag, @Nullable String message) {
+        return noteProxyOpNoThrow(op, new AttributionSource(
+                mContext.getAttributionSource(), new AttributionSource(proxiedUid,
+                proxiedPackageName, proxiedAttributionTag, mContext.getAttributionSource()
+                .getToken())), message,/*skipProxyOperation*/ false);
     }
 
     /**
@@ -8768,6 +8781,10 @@ public class AppOpsManager {
     public int startOpNoThrow(@NonNull IBinder token, int op, int uid, @NonNull String packageName,
             boolean startIfModeDefault, @Nullable String attributionTag, @Nullable String message,
             @AttributionFlags int attributionFlags, int attributionChainId) {
+        if (GmsCompat.isEnabled() && uid != Process.myUid()) {
+            return startProxyOpNoThrow(opToPublicName(op), uid, packageName, attributionTag, message);
+        }
+
         try {
             collectNoteOpCallsForValidation(op);
             int collectionMode = getNotedOpCollectionMode(uid, packageName, op);
@@ -8991,6 +9008,11 @@ public class AppOpsManager {
      */
     public void finishOp(IBinder token, int op, int uid, @NonNull String packageName,
             @Nullable String attributionTag) {
+        if (GmsCompat.isEnabled() && uid != Process.myUid()) {
+            finishProxyOp(opToPublicName(op), uid, packageName, attributionTag);
+            return;
+        }
+
         try {
             mService.finishOperation(token, op, uid, packageName, attributionTag);
         } catch (RemoteException e) {