--- a/core/java/android/app/ApplicationPackageManager.java
+++ b/core/java/android/app/ApplicationPackageManager.java
@@ -32,6 +32,7 @@ import android.annotation.Nullable;
 import android.annotation.StringRes;
 import android.annotation.UserIdInt;
 import android.annotation.XmlRes;
+import android.app.compat.gms.GmsCompat;
 import android.app.role.RoleManager;
 import android.compat.annotation.UnsupportedAppUsage;
 import android.content.ComponentName;
@@ -121,6 +122,9 @@ import android.util.Log;
 import com.android.internal.annotations.GuardedBy;
 import com.android.internal.annotations.Immutable;
 import com.android.internal.annotations.VisibleForTesting;
+import com.android.internal.gmscompat.GmsHooks;
+import com.android.internal.gmscompat.GmsInfo;
+import com.android.internal.gmscompat.PlayStoreHooks;
 import com.android.internal.os.SomeArgs;
 import com.android.internal.util.UserIcons;
 
@@ -217,6 +221,8 @@ public class ApplicationPackageManager extends PackageManager {
     @Override
     public PackageInfo getPackageInfo(VersionedPackage versionedPackage, int flags)
             throws NameNotFoundException {
+        flags = GmsHooks.filterPackageInfoFlags(flags);
+
         final int userId = getUserId();
         try {
             PackageInfo pi = mPM.getPackageInfoVersioned(versionedPackage,
@@ -233,6 +239,8 @@ public class ApplicationPackageManager extends PackageManager {
     @Override
     public PackageInfo getPackageInfoAsUser(String packageName, int flags, int userId)
             throws NameNotFoundException {
+        flags = GmsHooks.filterPackageInfoFlags(flags);
+
         PackageInfo pi =
                 getPackageInfoAsUserCached(
                         packageName,
@@ -452,6 +460,17 @@ public class ApplicationPackageManager extends PackageManager {
         if (ai == null) {
             throw new NameNotFoundException(packageName);
         }
+
+        if (GmsInfo.PACKAGE_GMS_CORE.equals(packageName)) {
+            // checked before accessing com.google.android.gms.phenotype content provider
+            // in com.google.android.libraries.phenotype.client
+            // .PhenotypeClientHelper#validateContentProvider() -> isGmsCorePreinstalled()
+            // PhenotypeFlags will always return their default values if these flags aren't set
+            if (GmsCompat.isGmsCore() || GmsCompat.isClientOfGmsCore()) {
+                ai.flags |= ApplicationInfo.FLAG_SYSTEM | ApplicationInfo.FLAG_UPDATED_SYSTEM_APP;
+            }
+        }
+
         return maybeAdjustApplicationInfo(ai);
     }
 
@@ -573,6 +592,10 @@ public class ApplicationPackageManager extends PackageManager {
     /** @hide */
     @Override
     public @NonNull List<SharedLibraryInfo> getSharedLibraries(int flags) {
+        if (GmsCompat.isEnabled()) {
+            // MATCH_ANY_USER requires privileged INTERACT_ACROSS_USERS permission
+            flags &= ~MATCH_ANY_USER;
+        }
         return getSharedLibrariesAsUser(flags, getUserId());
     }
 
@@ -761,6 +784,12 @@ public class ApplicationPackageManager extends PackageManager {
             if (Arrays.asList(featuresPixel6).contains(name)) return false;
         }
         if (Arrays.asList(featuresPixel).contains(name)) return true;
+        if (GmsCompat.isEnabled()) {
+            if (GmsHooks.isHiddenSystemFeature(name)) {
+                return false;
+            }
+        }
+
         return mHasSystemFeatureCache.query(new HasSystemFeatureQuery(name, version));
     }
 
@@ -1146,6 +1175,8 @@ public class ApplicationPackageManager extends PackageManager {
     @SuppressWarnings("unchecked")
     @Override
     public List<PackageInfo> getInstalledPackages(int flags) {
+        flags = GmsHooks.filterPackageInfoFlags(flags);
+
         return getInstalledPackagesAsUser(flags, getUserId());
     }
 
@@ -1868,6 +1899,10 @@ public class ApplicationPackageManager extends PackageManager {
 
     @Override
     public void addOnPermissionsChangeListener(OnPermissionsChangedListener listener) {
+        if (GmsCompat.isEnabled()) {
+            return;
+        }
+
         getPermissionManager().addOnPermissionsChangeListener(listener);
     }
 
@@ -2563,6 +2598,10 @@ public class ApplicationPackageManager extends PackageManager {
     @Override
     @UnsupportedAppUsage
     public void deletePackage(String packageName, IPackageDeleteObserver observer, int flags) {
+        if (GmsCompat.isPlayStore()) {
+            PlayStoreHooks.deletePackage(mContext, this, packageName, observer, flags);
+            return;
+        }
         deletePackageAsUser(packageName, observer, flags, getUserId());
     }
 
@@ -2609,6 +2648,10 @@ public class ApplicationPackageManager extends PackageManager {
     @Override
     public void freeStorageAndNotify(String volumeUuid, long idealStorageSize,
             IPackageDataObserver observer) {
+        if (GmsCompat.isPlayStore()) {
+            PlayStoreHooks.freeStorageAndNotify(mContext, volumeUuid, idealStorageSize, observer);
+            return;
+        }
         try {
             mPM.freeStorageAndNotify(volumeUuid, idealStorageSize, 0, observer);
         } catch (RemoteException e) {
@@ -2872,6 +2915,11 @@ public class ApplicationPackageManager extends PackageManager {
     @Override
     public void setApplicationEnabledSetting(String packageName,
                                              int newState, int flags) {
+        if (GmsCompat.isPlayStore()) {
+            PlayStoreHooks.setApplicationEnabledSetting(packageName, newState);
+            return;
+        }
+
         try {
             mPM.setApplicationEnabledSetting(packageName, newState, flags,
                     getUserId(), mContext.getOpPackageName());