--- a/core/java/com/android/internal/gmscompat/dynamite/server/IFileProxyService.aidl
+++ b/core/java/com/android/internal/gmscompat/dynamite/server/IFileProxyService.aidl
@@ -1,0 +1,5 @@
+package com.android.internal.gmscompat.dynamite.server;
+
+interface IFileProxyService {
+    ParcelFileDescriptor openFile(String path);
+}