… core/api/system-current.txt
… core/java/android/app/compat/gms/GmsCompat.java
… core/java/android/content/ContentResolver.java
… core/java/android/content/pm/PackageInstaller.java
… core/java/android/content/pm/PackageParser.java
… core/java/android/content/res/ApkAssets.java
… core/java/android/net/NetworkScoreManager.java
… core/java/android/os/Binder.java
… core/java/android/os/Build.java
… core/java/android/os/DropBoxManager.java
… core/java/android/os/Parcel.java
… core/java/android/os/UserHandle.java
… core/java/android/os/UserManager.java
… core/java/android/provider/ContactsContract.java
… core/java/android/provider/DeviceConfig.java
… core/java/android/provider/Settings.java
… core/java/com/android/internal/gmscompat/BinderRedirector.java
… core/java/com/android/internal/gmscompat/GmsCompatApp.java
… core/java/com/android/internal/gmscompat/GmsHooks.java
… core/java/com/android/internal/gmscompat/GmsInfo.java
… core/java/com/android/internal/gmscompat/GmsUserHooks.java
… core/java/com/android/internal/gmscompat/HybridBinder.java
… core/java/com/android/internal/gmscompat/IClientOfGmsCore2Gca.aidl
… core/java/com/android/internal/gmscompat/IGms2Gca.aidl
… core/java/com/android/internal/gmscompat/PlayStoreHooks.java
… core/java/com/android/internal/gmscompat/dynamite/GmsDynamiteClientHooks.java
… core/java/com/android/internal/gmscompat/dynamite/server/FileProxyService.java
… core/res/AndroidManifest.xml
… location/java/android/location/LocationManager.java
… packages/PackageInstaller/src/com/android/packageinstaller/PackageInstallerActivity.java