commit 1023690c3f2515bcc540c945c2576e2607a35bf6
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Tue Apr 19 14:31:41 2022 +0300

    gmscompat: conditionally stub out BluetoothAdapter.get{Address,Name}()

commit aa97c368f015e327692fe1a01d6edc7c43bbfd04
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Mon Apr 18 15:08:14 2022 +0300

    gmscompat: simplify initialization

commit ca20e5770bf43e80bbbaabca9c996d64b38b8e31
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Mon Apr 18 14:40:32 2022 +0300

    gmscompat: initialize GmsCompat before execution of the GMS code

    Application.attach() calls attachBaseContext() that is overridden in
    GMS Core, which means that GMS Core performs early init when GmsCompat
    is not enabled and fails to do so in some cases.

    Instantiation of the custom Application subclass isn't overridden
    by GMS currently, but could be in the future.

commit 522456bf2d20c933df5425db30136f800c36e6b8
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Mon Apr 18 14:32:04 2022 +0300

    gmscompat: add an option to match disabled packages in isGmsApp()

commit 42effdc33e8dbf3a4429895dcacd85824187dd24
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sun Apr 17 17:37:35 2022 +0300

    gmscompat: don't hide the BluetoothManager

    Hiding it leads to FIDO2 service not working without the "Nearby
    devices" permission, which is a regression.

commit 2aacfca26b2293b76b9e12186f7afaa94a2d3679
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Thu Apr 14 17:05:26 2022 +0300

    gmscompat: improve the package uninstallation shim

commit b8c216851f24add5f87f950839d04b8c5c282f20
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Thu Apr 14 17:02:29 2022 +0300

    gmscompat: add queryStatsForPackage() shim

commit 89ef73f5887d8ff4809d190ff46521c3a1aa7d3a
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Thu Apr 14 17:01:53 2022 +0300

    gmscompat: improve work profile support

commit b8c0df3b21570603d07dfa3c36178a1ad7c3dc44
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Thu Apr 14 17:00:17 2022 +0300

    gmscompat: skip privileged setBackgroundActivityStartsAllowed() method

commit 76ef08fdf7b58c917c837ebf9e4726b1a10f0401
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Wed Apr 13 22:16:41 2022 +0300

    gmscompat: skip privileged DownloadRequest property

commit 6c68e0cada9387673259629e271c19cbc5c5ad22
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sun Apr 10 18:57:28 2022 +0300

    gmscompat: consider disabled GMS components to be absent

commit 7e33e82d4e80e343397d9ff51252c516f3d8e9d9
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sun Apr 10 16:29:00 2022 +0300

    gmscompat: filter out privileged Downloads.Impl.VISIBILITY_HIDDEN

commit 13b9ede18702224909fe0eff3ac972454f7c9112
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sun Apr 10 15:34:50 2022 +0300

    gmscompat: handle background Activity starts

commit 7e1f8705d3f020cec8a74f761c70d8f3531a88b2
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sun Apr 10 15:28:18 2022 +0300

    gmscompat: add BluetoothDevice.getMetadata() shim

commit f03a44ef17d1972cea2747a4dec1d5f1cad29b95
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sat Apr 9 16:03:00 2022 +0300

    gmscompat: notify GmsCompatApp when a notable interface is acquired

commit 9b9edfa78c4d6dd61400456135a9d24d94209be9
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sat Apr 9 16:01:55 2022 +0300

    gmscompat: export GmsCompat.isGmsApp(pkgName, userId) method

    Needed by PermissionController and GmsCompat app.

commit a9208fae82b341eeb249d597249014a23f1b6a60
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sat Apr 9 16:00:15 2022 +0300

    gmscompat: notify user when extra permission is needed to install an app

commit 4253b53dfdd10ea2ce7717e8eacfca3016ff5256
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Thu Apr 7 17:53:27 2022 +0300

    gmscompat: conditionally stub out getSimContacts()

commit 98d77a3e93b317bc33c6223b6665d7f836d8b5fd
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Thu Apr 7 17:52:45 2022 +0300

    gmscompat: conditionally hide BluetoothManager

    There are multiple GMS Core services that crash when "Nearby devices"
    permission is revoked, but move to fallback paths when
    BluetoothManager is null or relevant system features are absent.

commit 3b0ea425d6a086aa9df15e7550471d2d77f8be73
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Wed Apr 6 11:36:48 2022 +0300

    gmscompat: use app Context in isClientOfGmsCore()

    Context that is passed from the outside could potentially lead
    to a wrong result.

commit c7547346222fe5cb57f57f8f5f807db879268a0f
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Wed Apr 6 11:34:12 2022 +0300

    gmscompat: fix false positive isClientOfGmsCore() result

    isClientOfGmsCore() can be called indirectly by the framework code
    before maybeInit() is called. At that point, isGmsCore() returns
    false in GMS Core process, because isGmsCore variable is at its default
    "false" value. This leads to GMS Core trying to use client compat code
    for Dynamite module support and crashing.

commit b91c2666dc02c44cf088fd9030b69de817ecce69
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Tue Apr 5 15:18:45 2022 +0300

    gmscompat: add missing FLAG_ACTIVITY_NEW_TASK

commit 326f68e7f627fbdc8aff40c86c679d3cb9d3dc22
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Mon Apr 4 18:26:45 2022 +0300

    gmscompat: improve PlayStoreHooks

    Issue "pending user action" notifications from GmsCompatApp to make
    the code simpler and more reliable (Play Store may cancel
    its notifications in multiple ways).

    Also, check PackageInstaller confirmation intents in Activity.onResume()
    to make sure that there are no blocked PackageInstaller sessions.

commit a671c82d2bad552db14c95818c9cce99fc8b508e
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Mon Apr 4 18:14:22 2022 +0300

    gmscompat: use AIDL instead of ContentProvider.call()

commit 151f40dcb60cf154eb83a48e549df2c4cf1d874e
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Mon Apr 4 18:10:19 2022 +0300

    gmscompat: additional shims

commit 79ae9544cb77f170edece97c8361a23c508d17ae
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Tue Mar 29 10:45:29 2022 +0300

    gmscompat: use a proper name to refer to the Play services app

commit 023e32e6a40999c374bbbe049a590705188a66c2
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Tue Mar 15 12:45:38 2022 +0200

    gmscompat: improve Dynamite module support

    - support DelegateLastClassLoaders that don't define librarySearchPath
    (eg DynamiteLoader.apk) by checking "dexPath" too
    - make the class loader path check faster now that there's two of them
    - enable Dynamite modules in Play Store processes
    - use the "/proc/self/fd" fd passing approach for opening the dex files
    - don't track the module loading state, cache file descriptor lookups
    instead. Previous approach stopped working when module loading sequence
    changed (extra File.lastModified() check was added)
    - use a correct zipFileSeparator ("!/", not "!")
    - support ApkAssets that use an AssetsProvider
    - use "/proc/self/fd" path for File.lastModified() hook
    instead of an extra IPC to Play services

commit 423068be2d7ae72020c1c144f52c74fa7205faf0
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Fri Mar 25 17:30:00 2022 +0200

    gmscompat: defer GMS client checks

    Removes unconditional PackageManager call at startup of each app process
    and ensures correct behavior of GMS clients that are launched before
    GMS is installed (Dynamite clients would crash, BinderRedirector callers
    wouldn't get HybridBinder, etc)

commit aa48504da62d45056a65e7ad486a144396ec4ab6
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Mon Mar 14 20:29:08 2022 +0200

    gmscompat: simplify access to the FileProxyService

commit 608cb2590c20e7d65ee7fa63ef82cb95145c3e19
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Mon Mar 14 20:22:31 2022 +0200

    gmscompat: improve FileProxyService

    - disallow opening of directories
    - restrict file type to ".apk"
    - don't forward exceptions to minimize the information leaks
    - don't check accessibility of "app_chimera/m/" components that
    are known to be accessible
    - optimize the path splitting operations
    - remove the getLastModified() method that is no longer needed

commit e9b61002bbac471cf752892d4161cd223a4acaed
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Tue Mar 15 12:49:03 2022 +0200

    gmscompat: expand PhenotypeFlags workaround to all GMS clients

commit 4a05b591ba3a0987fa25fb69ee01a3ea99809050
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Mon Mar 7 16:22:16 2022 +0200

    gmscompat: hide privileged FontManager service

commit c96859fca04621a101fbc60f10d2e59cf83149df
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sun Mar 6 22:38:07 2022 +0200

    gmscompat: fix location service regression

commit 04b94558dbcde70d18e75a6cc61d81badd2d52b8
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sun Mar 6 22:30:38 2022 +0200

    gmscompat: document PhenotypeFlag provider workaround

commit 2747ddc55a841ec2ae159357e7986eb50d4dad95
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Wed Feb 2 15:47:53 2022 +0200

    gmscompat: improve userId-related shims

    Current approach is fragile: GMS always obtains UserHandle of the primary ("Owner") user and
    tries to use it, which is a privileged operation if GMS is installed in the non-primary user profile.
    It's worked around by unconditionally replacing UserHandle with the current one in all methods that
    GMS calls, which will break each time new GMS version makes use of a new UserHandle-accepting method.

    With the new approach, GMS gets access only to the current UserHandle which is made
    to look like a UserHandle of the primary user.

commit 9d83ecdf111a261f30dcdc6978117d9277dd1aa7
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Tue Feb 1 17:48:42 2022 +0200

    gmscompat: use correct "isApplicationUid" check

commit 0c2fa410c0e14f03bef86a9d4636c169ad797750
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Wed Feb 2 12:18:25 2022 +0200

    gmscompat: defer isGmsApp() check

commit 849f3cd0cefcc6d53c7f32eba853d82ec09c7636
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sat Feb 5 11:51:14 2022 +0200

    Revert "gmscompat: don't check whether Google account removal is allowed"

    Not needed anymore.

commit 67e7cb2b08cae8ac213dc3d327fa3efe6fc93055
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sat Feb 5 11:49:40 2022 +0200

    gmscompat: account removal workaround

commit cb0bb92fc495b84e59c33edd661106f4e6ce566d
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sat Mar 5 19:20:08 2022 +0200

    gmscompat: mark public ModuleLoadState fields final

commit f31ad2afc27832cc2eba6af652c12651c9663d93
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sat Mar 5 19:18:02 2022 +0200

    gmscompat: optimize file path check

    getAbsolutePath() performs needless checks

commit 1fe36dcc1e4c60d2e35c91aa23fa3dc58c6a8e7a
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sat Mar 5 19:12:51 2022 +0200

    gmscompat: fix TOCTOU race

commit 557248abfcc7f2594b6ff0d13e87583a8f83b7e0
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sat Mar 5 19:01:32 2022 +0200

    gmscompat: simplify AppIntegrityManager shim

commit f2a2f7d804ab256d5bed2d12fe30d8d0aaf37d35
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Sat Feb 26 22:25:48 2022 +0200

    gmscompat: stop Play Store from trying to auto-update some apps

commit cfc13eec503b08875b1b782085db9e891f94934f
Author: Danny Lin <danny@kdrag0n.dev>
Date:   Sat Feb 19 23:56:40 2022 -0800

    gmscompat: Improve GMS package info exception handling

    Change-Id: Ibb74bd4de8233284d42341cb3a732cdc9107c9d7

commit 7da8ebc6569dcb7bdbe59fc5140ddc1556f929bd
Author: Danny Lin <danny@kdrag0n.dev>
Date:   Thu Feb 17 21:59:01 2022 -0800

    gmscompat: Minor code style fixes

    Change-Id: I88b5e87ab429c409bca4226f7566c56eb87af8ff

commit 48836d8e20551c73b2ae115409321f87faef0fdf
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Thu Feb 17 22:20:38 2022 +0200

    gmscompat: hide UWB system feature

commit 2d758a1b872512afc0fe33add292eb97c082c8fe
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Tue Feb 15 19:27:06 2022 +0200

    gmscompat: compatibility with updated Dynamite module loading sequence

    There's now additional File.lastModified() checks that happen outside
    the main module loading sequence.

commit 128fb3745a341b19e7057a9ba33e6814cb81f8b1
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Tue Feb 15 18:57:25 2022 +0200

    gmscompat: enable on-demand Dynamite modules

commit 47ca582bf117b64d8c68557197ab0ad0a1d9435e
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Wed Feb 9 15:30:10 2022 +0200

    gmscompat: perform binder redirection check in IGmsCallbacks interface

commit ad064e3e38e9b8ea320d22db7fec498a4b6dd5eb
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Wed Feb 9 15:25:29 2022 +0200

    gmscompat: add BinderRedirector

commit 94678363efa29d5f6c9b5f9c99893a280845c209
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Wed Feb 9 15:06:30 2022 +0200

    gmscompat: add HybridBinder

commit 1854611fd2fccba4e91bb63ed11f0863a974d384
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Mon Jan 31 19:55:48 2022 +0200

    gmscompat: simplify initialization

commit 1152abf4441c749f05b9cb9260c157a181bc544c
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Mon Jan 24 18:54:25 2022 +0200

    gmscompat: make connection to GmsCompatApp unbreakable

commit 4b52810dc052d01597d6756f440da6ec9367cb1b
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Fri Jan 21 21:44:14 2022 +0200

    gmscompat: don't make all services foreground

    Instead, rely on an external foreground app to bind to GMS, which
    allows GMS to start regular services.

commit 1c075ea53be9832c12ecf36eac9e910cb4744313
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Fri Jan 21 21:40:07 2022 +0200

    gmscompat: improve location services compatibility

commit 4cad08135f3f6661044bbefdce72bf91d708bead
Author: Daniel Micay <danielmicay@gmail.com>
Date:   Sat Jan 29 01:23:35 2022 -0500

    Revert "gmscompat: set foregroundServiceType of location services"

    This reverts commit 8ec949adf42a16360f23e2986c05316dbf99bf16.

commit 44c4e10b42fa55967d43115f9cff037d2468ff11
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Thu Jan 13 21:38:26 2022 +0200

    gmscompat: don't stub out PackageManager.getSharedLibraries()

    Play Store uses list of shared libraries to filter out apps with missing dependencies.

commit f2e676248f75b6b2b66e1c5ec33f5403fd5cee4f
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Wed Jan 12 21:52:27 2022 +0200

    gmscompat: enable Play {Asset, Feature} Delivery

    Play Store uses a similar getRunningAppProcesses()-based "is client foreground"
    check as Play Services does, so enable the same workaround for it as well.

commit 841169c0f6ca3bd3efc111688755e64af423381e
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Mon Jan 10 10:21:40 2022 +0200

    gmscompat: don't check whether Google account removal is allowed

    When user profile is protected by PIN or password ("keyguard is secure"),
    GMS tries to access privileged APIs when asked whether account removal is allowed,
    which leads to a negative reply and unremovable account (unless GMS itself is removed).

commit 27c723b4f74204da9ace6cf55a207cce1f16eae3
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Mon Jan 10 10:05:02 2022 +0200

    gmscompat: improve account sign-in UI

    GMS tries to access BackupService after sign-in, but doesn't have the permission
    to do so. This manifests itself in the UI as a never-ending spinner.

commit e4d51b9c80a8732c7d28e3e262d001159d828178
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Tue Dec 14 17:33:44 2021 +0200

    gmscompat: add recently bound pids to getRunningAppProcesses() result

    Play Games Services relies on ActivityManager.getRunningAppProcesses()
    to figure out whether its client is running.
    Unprivileged apps see only their own processes.

commit a2e9d682998468b9b4c0770150c641642688b401
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Wed Dec 8 19:54:15 2021 +0200

    gmscompat: set foregroundServiceType of location services

    Without it, location access gets throttled after a few minutes.

commit 2a94b95caf21737daf325ed1015bd46cb7f60809
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Thu Nov 25 21:56:12 2021 +0200

    gmscompat: System.exit(1) if package uninstallation is declined

commit f96dbb64ca9c14078e18b4e68b9209ca541c543e
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Thu Nov 25 21:53:44 2021 +0200

    gmscompat: use a reliable "is app foreground" check

    Previous check returned a false positive shortly after activity was paused,
    which led to startActivity() silently failing.

commit d145d2aa9886c50eafda8b53f1d6add33c16d55e
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Thu Nov 25 21:46:42 2021 +0200

    gmscompat: make sure PackageInstaller UI returns a result

commit 3e494d0c8e8a8e4bbec7e793ad67cc43f5b18b02
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Tue Nov 23 21:02:26 2021 +0200

    gmscompat: fix push notifications in secondary user profiles

    UserManager.isUserUnlocked(userId) returns whether a user exited the Direct Boot mode.
    This method requires a restricted permission if userId parameter differs from the current userId.
    GMS called this method with userId 0 in both primary (userId == 0) and secondary (userId != 0)
    profiles. userId was always 0 likely because of the stubbed out
    UserManager.getUserSerialNumber(userId),
    which always returned 0 as a workaround to a different issue.

commit 42daf020655559887ebf50c8d9b372aafe3fe831
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Mon Nov 22 17:56:16 2021 +0200

    gmscompat: mark confirmation notification from Play Store as ongoing

    Play Store will wait for confirmation indefinitely if notification is dismissed.
    Ongoing notifications can't be dismissed (swiped away) by the user.

commit 7adbe10ce2c2807eab6604efee30476b230d1f55
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Fri Nov 19 13:48:41 2021 +0200

    gmscompat: compatibility with Play Store self-update

commit 5a184d7533f9de7c25bbb8b758cc5db1ea903aa2
Author: Dmitry Muhomor <muhomor.dmitry@gmail.com>
Date:   Tue Nov 16 18:54:30 2021 +0200

    gmscompat: improve Play Store compatibility

    Implements package (un)installation confirmation handling and
    enables unattended app updates.

commit 7632b0fe63cf73acc77a3ca4bfcad8e059283d93
Author: Daniel Micay <danielmicay@gmail.com>
Date:   Sat Nov 13 18:38:39 2021 -0500

    gmscompat: use new API for long-running AppOps

    Android 12 added startProxyOpNoThrow and finishProxyOp so we no longer
…